remove namespace Performance from test
