+++ sdk/identity/azure-identity/test/perf/CMakeLists.txt
@@ -9,7 +9,7 @@ set(CMAKE_CXX_STANDARD_REQUIRED True)
 
 set(
   AZURE_IDENTITY_PERF_TEST_HEADER
-  inc/azure/identity/test/perf/secret_credential.hpp
+  inc/azure/identity/test/secret_credential.hpp
 )
 
 set(
