+++ sdk/identity/azure-identity/test/perf/src/main.cpp
@@ -3,14 +3,14 @@
 
 #include <azure/perf.hpp>
 
-#include "azure/identity/test/perf/secret_credential.hpp"
+#include "azure/identity/test/secret_credential.hpp"
 
 int main(int argc, char** argv)
 {
 
   // Create the test list
   std::vector<Azure::Perf::TestMetadata> tests{
-      Azure::Identity::Test::Performance::SecretCredentialTest::GetTestMetadata()};
+      Azure::Identity::Test::SecretCredentialTest::GetTestMetadata()};
 
   Azure::Perf::Program::Run(Azure::Core::GetApplicationContext(), tests, argc, argv);
