+++ sdk/keyvault/azure-security-keyvault-keys/test/perf/src/main.cpp
@@ -3,14 +3,14 @@
 
 #include <azure/perf.hpp>
 
-#include "azure/keyvault/keys/test/perf/get_key.hpp"
+#include "azure/keyvault/keys/test/get_key.hpp"
 
 int main(int argc, char** argv)
 {
 
   // Create the test list
   std::vector<Azure::Perf::TestMetadata> tests{
-      Azure::Security::KeyVault::Keys::Test::Performance::GetKey::GetTestMetadata()};
+      Azure::Security::KeyVault::Keys::Test::GetKey::GetTestMetadata()};
 
   Azure::Perf::Program::Run(Azure::Core::GetApplicationContext(), tests, argc, argv);
