+++ sdk/storage/azure-storage-blobs/test/perf/src/main.cpp
@@ -3,14 +3,14 @@
 
 #include <azure/perf.hpp>
 
-#include "azure/storage/blobs/test/perf/download_blob.hpp"
+#include "azure/storage/blobs/test/download_blob.hpp"
 
 int main(int argc, char** argv)
 {
 
   // Create the test list
   std::vector<Azure::Perf::TestMetadata> tests{
-      Azure::Storage::Blobs::Test::Performance::DownloadBlob::GetTestMetadata()};
+      Azure::Storage::Blobs::Test::DownloadBlob::GetTestMetadata()};
 
   Azure::Perf::Program::Run(Azure::Core::GetApplicationContext(), tests, argc, argv);
